Collapse most of graph and text unit-based report generation
