--- graphrag/index/operations/summarize_communities/community_reports_extractor/utils.py
@@ -1,15 +1,0 @@
-# Copyright (c) 2024 Microsoft Corporation.
-# Licensed under the MIT License
-
-"""A module containing community report generation utilities."""
-
-import pandas as pd
-
-import graphrag.index.operations.summarize_communities.community_reports_extractor.schemas as schemas
-
-
-def get_levels(df: pd.DataFrame, level_column: str = schemas.NODE_LEVEL) -> list[int]:
-    """Get the levels of the communities."""
-    levels = df[level_column].dropna().unique()
-    levels = [int(lvl) for lvl in levels if lvl != -1]
-    return sorted(levels, reverse=True)
