+++ graphrag/index/operations/summarize_communities/graph_context/__init__.py
@@ -1,4 +1,4 @@
 # Copyright (c) 2024 Microsoft Corporation.
 # Licensed under the MIT License
 
-"""The Indexing Engine community reports package root."""
+"""Package of context builders for graph-based reports."""
