+++ graphrag/index/operations/summarize_communities/utils.py
@@ -1,16 +1,20 @@
 # Copyright (c) 2024 Microsoft Corporation.
 # Licensed under the MIT License
 
-"""A module containing create_graph, _get_node_attributes, _get_edge_attributes and _get_attribute_column_mapping methods definition."""
+"""A module containing community report generation utilities."""
 
-import logging
 from itertools import pairwise
 
 import pandas as pd
 
 import graphrag.index.operations.summarize_communities.community_reports_extractor.schemas as schemas
 
-log = logging.getLogger(__name__)
+
+def get_levels(df: pd.DataFrame, level_column: str = schemas.NODE_LEVEL) -> list[int]:
+    """Get the levels of the communities."""
+    levels = df[level_column].dropna().unique()
+    levels = [int(lvl) for lvl in levels if lvl != -1]
+    return sorted(levels, reverse=True)
 
 
 def restore_community_hierarchy(
