+++ graphrag/index/workflows/factory.py
@@ -30,8 +30,8 @@ def _get_workflows_list(
                 "create_base_text_units",
                 "create_final_documents",
                 "extract_graph",
-                "create_communities",
                 *(["extract_covariates"] if config.claim_extraction.enabled else []),
+                "create_communities",
                 "create_final_text_units",
                 "create_community_reports",
                 "generate_text_embeddings",
