+++ graphrag/query/indexer_adapters.py
@@ -12,7 +12,9 @@ from typing import cast
 import pandas as pd
 
 from graphrag.config.models.graph_rag_config import GraphRagConfig
-from graphrag.index.operations.summarize_communities import restore_community_hierarchy
+from graphrag.index.operations.summarize_communities.utils import (
+    restore_community_hierarchy,
+)
 from graphrag.model.community import Community
 from graphrag.model.community_report import CommunityReport
 from graphrag.model.covariate import Covariate
